tools/requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.1
websockets>=14.0
